[Refactor] Aside 폴더 구조재정리(#79)
