+++ src/components/aside/Aside.jsx
@@ -1,19 +1,19 @@
 import { useState } from 'react';
 import './Aside.css';
 import { useCookies } from 'react-cookie';
-import tagIcon from '../../../assets/icons/tag-icon.svg';
-import chatIcon from '../../../assets/icons/chat-icon.svg';
-import personIcon from '../../../assets/icons/person-icon.svg';
-import fileIcon from '../../../assets/icons/file-icon.svg';
-import writingIcon from '../../../assets/icons/writing-icon.svg';
-import mobileAsideIcon from '../../../assets/icons/mobile-aside-icon.svg';
+import tagIcon from '../../assets/icons/tag-icon.svg';
+import chatIcon from '../../assets/icons/chat-icon.svg';
+import personIcon from '../../assets/icons/person-icon.svg';
+import fileIcon from '../../assets/icons/file-icon.svg';
+import writingIcon from '../../assets/icons/writing-icon.svg';
+import mobileAsideIcon from '../../assets/icons/mobile-aside-icon.svg';
 import { Link, useNavigate } from 'react-router-dom';
-import useCurrentPath from '../../../hooks/useCurrentPath';
-import useLastCommentStore from '../../../store/useLastCommenStore';
-import useShowExtend from '../AsideMenu/hooks/useShowExtend';
-import WritingMenu from '../WritingMenu/WritingMenu';
-import useIsMobile from '../../../hooks/useIsMobile';
-import FileMenu from '../FileMenu/FileMenu';
+import useCurrentPath from '../../hooks/useCurrentPath';
+import useLastCommentStore from '../../store/useLastCommenStore';
+import useShowExtend from './hooks/useShowExtend';
+import WritingMenu from './WritingMenu/WritingMenu';
+import useIsMobile from '../../hooks/useIsMobile';
+import FileMenu from './FileMenu/FileMenu';
 
 function Aside() {
   const [cookies] = useCookies(['studentId']);
